gc - trying to get tests working that parse the procfile

+++ src/_smoke_test.js
@@ -17,6 +17,7 @@
         child_process = require("child_process"),
         http = require("http"),
         PORT_NUM = "5000",
+        fs = require("fs"),
         child;
 
 //    exports.setUp = function (done) {
@@ -28,14 +29,20 @@
     };
 
     exports.tearDown = function (done) {
-        child.on("exit", function (code, signal) {
-            done();
-        });
-        child.kill();
+        if (!child) {
+            console.log("Child is null");
+            return;
+        } else {
+            console.log("Child is NOT null");
+            child.on("exit", function (code, signal) {
+                done();
+            });
+            child.kill();
+        }
     };
 
     // TODO: factor out common server name
-    exports._test_canGetHomePage = function (test) {
+    exports.test_canGetHomePage = function (test) {
         httpGet("http://localhost:" + PORT_NUM, function (response, receivedData) {
             var foundHomePage = (receivedData.indexOf("Welcome to WikiPaint!") !== -1);
             test.ok(foundHomePage, "home page should have contained 'Welcome to WikiPaint!'.");
@@ -56,8 +63,11 @@
         console.log("command line: " + commandLine);
         console.log("command line[0]: " + commandLine[0]);
         console.log("command line[1]: " + commandLine[1]);
+        console.log("command line[2]: " + commandLine[2]);
 
-        child = child_process.spawn(commandLine[0], commandLine[1]);
+        console.log("starting process");
+        child = child_process.spawn(commandLine[0], commandLine[1], commandLine[2]);
+
         child.stdout.setEncoding("utf8");
         child.stdout.on("data", function (chunk) {
             if (chunk.trim() === "Server started") {
@@ -65,10 +75,19 @@
             }
 
         });
+        console.log("started process");
     }
 
     function parseProcFile() {
-        return ["node", "src/server/wikipaint.js " + PORT_NUM];
+        var procFile = fs.readFileSync("Procfile", "UTF8"),
+            matches = "";
+
+        console.log(procFile);
+        matches = procFile.match(/^web:(\S)?(\S+\s+)?/); // matches 'web: one two three'
+
+        console.log("Matches: " + matches.join(", "));
+        //return procFile;
+        return ["node", "src/server/wikipaint.js", "8081"];
     }
 
     function httpGet(url, callback) {
